requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
anyio==4.0.0
    # via httpcore
black==23.9.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   black
    #   pip-tools
exceptiongroup==1.1.3
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
mypy==1.5.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.1
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.10.0
    # via black
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
ruff==0.0.291
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
toml==0.10.2
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pip-tools
    #   pyproject-hooks
tqdm==4.66.1
    # via -r requirements.in
types-toml==0.10.8.7
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   black
    #   mypy
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
